webui: move Guest Shares into the Sharing page as a tab
Guest Shares was under the Storage nav group, but it's conceptually a form
of sharing. Move it to be a peer tab on the Sharing page alongside SMB /
NFS / iSCSI / NVMe-oF, where the "sharing" mental model points, and drop the
separate Storage nav entry.

- New GuestSharesPanel.svelte holds the management table (moved verbatim
  from the old /shares page — list, Show-removed toggle, sortable columns,
  revoke/remove).
- Sharing page gains a "Guest Shares" tab. The cross-protocol Create-Share
  wizard and the protocol status dot are protocol-only, so they're hidden on
  the guest tab (guest links are created from the Files page). `shareProtocol`
  is now typed to the four protocols (ShareProto) so the wizard can't be
  pointed at 'guest'.
- /shares is kept as a redirect to /sharing#guest so existing links and the
  Files-page "Guest Shares" link still resolve.

Pure WebUI. svelte-check + build + vitest (144) all green.
